SECURITY: Fix XSS in breadcrumbs and status styles, fix note encoding
- Breadcrumb.php: Cast fromId to (int) to prevent reflected XSS via
  attribute injection in href attributes
- JobSummaryView.php: Escape statusStyle with htmlOut() before
  inserting into style attribute (matches JobListView pattern)
- SearchListView.php: Same fix for search status style
- NoteController.php: Remove htmlspecialchars() from update() — was
  causing progressive double-encoding on every save. add() correctly
  stores raw text; escaping belongs in the view layer.

+++ Libs/Breadcrumb.php
@@ -41,7 +41,7 @@ class Breadcrumb
     {
         $from = Tools::param('from');
         $q = Tools::param('q');
-        $fromId = Tools::param('fromId');
+        $fromId = (int) Tools::param('fromId');
         $fromName = Tools::param('fromName');
 
         $listPages = [
